Service, DTO and Controller fix

+++ src/main/java/cz/uun/companydatabase/service/CompanyService.java
@@ -8,6 +8,8 @@ import cz.uun.companydatabase.dtoin.CompanyUpdateDtoIn;
 import cz.uun.companydatabase.entity.Company;
 import cz.uun.companydatabase.repository.CompanyRepository;
 
+import java.util.Optional;
+
 @Service
 public class CompanyService {
     
@@ -15,25 +17,39 @@ public class CompanyService {
     private CompanyRepository companyRepository;
 
     public Company createCompany(CompanyCreateDtoIn dtoIn) {
-        Company company = companyRepository.save(new Company());
-        return company;
+        Company company = new Company();
+        company.setName(dtoIn.getName());
+        company.setAddress(dtoIn.getAddress());
+        company.setPhoneNumber(dtoIn.getPhoneNumber());
+        company.setIco(dtoIn.getIco());
+        return companyRepository.save(company);
     }
-    //TODO: Fix
+
     public Company getCompanyByIco(String ico) {
-        return companyRepository.findByIco(ico);
+        Optional<Company> companyOptional = companyRepository.findByIco(ico);
+        return companyOptional.orElse(null);
     }
 
     public Company deleteCompanyByIco(String ico) {
-        Company company = companyRepository.findByIco(ico);
-        companyRepository.delete(company);
-        return company;
+        Optional<Company> companyOptional = companyRepository.findByIco(ico);
+        if (companyOptional.isPresent()) {
+            Company company = companyOptional.get();
+            companyRepository.delete(company);
+            return company;
+        }
+        return null;
     }
 
     public Company updateCompanyByIco(CompanyUpdateDtoIn dtoIn) {
-        Company company = companyRepository.findByIco(dtoIn.getIco());
-        company.setIco(dtoIn.getIco());
-        company.setName(dtoIn.getName());
-        companyRepository.save(company);
-        return company;
+        Optional<Company> companyOptional = companyRepository.findByIco(dtoIn.getIco());
+        if (companyOptional.isPresent()) {
+            Company company = companyOptional.get();
+            company.setName(dtoIn.getName());
+            company.setAddress(dtoIn.getAddress());
+            company.setPhoneNumber(dtoIn.getPhoneNumber());
+            company.setIco(dtoIn.getIco());
+            return companyRepository.save(company);
+        }
+        return null;
     }
 }
